requirements.txt
ffmpeg-python
numpy
scipy
matplotlib
colour-science
ruamel.yaml
cerberus
tqdm